revert(chat): drop the no-reply composer mode, keep the append fix
Reverts the "send without asking for a reply" feature from #369 — the
composer toggle, `ChatNotifier.tryAppendMessage` / the `generateReply`
flag, the docs section and the `chat_no_reply_hint` string all go back to
the way they were.

What stays is the bug the feature uncovered: `_sendMessage` picked the
accepted assistant variation by scanning the history backwards, so with a
trailing user message it named an older variation than the one
`ChatRepo.appendUserMessageAndAcceptCurrentVariation` accepts ("a user
send accepts only the assistant immediately before it"). The repo refused
the guarded write, the append was silently dropped, and the message the
user had just sent disappeared from the chat. It now takes the trailing
message, and only when that message is an assistant.

Reachable without any new UI: send, press Stop before the first chunk
arrives (which leaves the chat ending on the user message), then send
again — the second message was lost.

- `test/send_after_aborted_send_test.dart` covers exactly that sequence;
  it fails on the old scan and passes on the fix.

Verified with `flutter analyze --no-fatal-infos --no-fatal-warnings` (28
issues, all pre-existing) and a full `flutter test` run (3671 passed).
